+++ OpenGLPuzzleCubeProject/src/Engine/Mesh.cpp
@@ -567,13 +567,13 @@ namespace Mesh {
 		if (!currentLevel.meshList.empty()) {
 			std::vector<int> removeList;
 			for (auto itr = currentLevel.meshList.begin(); itr != currentLevel.meshList.end(); itr++) {
-				for (int i = itr->second->beginMaterial; i < itr->second->endMaterial; i++) {
+				for (size_t i = itr->second->beginMaterial; i < itr->second->endMaterial; i++) {
 					removeList.push_back(i);
 				}
 			}
 			std::sort(removeList.begin(), removeList.end(), [&](int a, int b) {return a > b; });
 
-			for (int i = 0; i < removeList.size(); i++) {
+			for (size_t i = 0; i < removeList.size(); i++) {
 
 				materialList.erase(materialList.begin() + removeList[i]);
 
@@ -597,13 +597,13 @@ namespace Mesh {
 		//現在のレベルが管理しているマテリアルリストの範囲内を削除する
 		std::vector<int> removeList;
 		for (auto itr = currentLevel.meshList.begin(); itr != currentLevel.meshList.end(); itr++) {
-			for (int i = itr->second->beginMaterial; i < itr->second->endMaterial;i++) {
+			for (size_t i = itr->second->beginMaterial; i < itr->second->endMaterial;i++) {
 				removeList.push_back(i);
 			}
 		}
 		std::sort(removeList.begin(), removeList.end(), [&](int a, int b) {return a > b; });
 
-		for (int i = 0; i < removeList.size(); i++) {
+		for (size_t i = 0; i < removeList.size(); i++) {
 
 			materialList.erase(materialList.begin() + removeList[i]);
